requirements.txt
# ──────────────────────────────────────────────────────
# requirements.txt — RAG Mutual Fund FAQ Chatbot
# ──────────────────────────────────────────────────────
# Pinned versions for reproducible installs on
# Streamlit Community Cloud.  Last updated: 2026-02-26
# ──────────────────────────────────────────────────────

# LangChain — text splitting
langchain==0.3.19
langchain-community==0.3.17

# Vector database — Qdrant (in-memory mode)
qdrant-client==1.13.3

# Embeddings — OpenAI
openai==1.63.2

# LLM — Groq (Llama 3.1 8B)
groq==0.18.0

# Reranker — cross-encoder
sentence-transformers==3.4.1

# PDF extraction (ingestion phase)
pdfplumber==0.11.6

# Web UI
streamlit==1.42.2

# Environment variable loading
python-dotenv==1.0.1

# Utilities (transitive but pinned for stability)
tiktoken==0.8.0
numpy==1.26.4
